requirements.txt
pandas==2.2.3
python-telegram-bot==20.8
requests==2.32.3
tzdata==2025.2
websockets==12.0
